+++ htdocs/synopsistools/git_maj_version.php
@@ -7,6 +7,6 @@ require_once DOL_DOCUMENT_ROOT . '/bimpcore/Bimp_Lib.php';
 
 sleep(10);
 
-$version = BimpCore::getConf('git_version', 1)+1;
+$version = (int) BimpCore::getConf('git_version', 1)+1;
 echo 'version '.$version;
 BimpCore::setConf('git_version', $version);
